# client/web/openapi.json
{
  "components": {
    "schemas": {
...
      "Bot": {
        "properties": {
          "archived_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archived At"
          },
          "color": {
            "title": "Color",
            "type": "string"
          },
          "computer_mode": {
            "enum": [
              "team",
              "dedicated"
            ],
            "title": "Computer Mode",
            "type": "string"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "cursor_agent_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cursor Agent Id"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "id": {
            "minLength": 1,
            "title": "Id",
            "type": "string"
          },
          "instructions": {
            "title": "Instructions",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "notify_on_finish": {
            "title": "Notify On Finish",
            "type": "boolean"
          },
          "parent_bot_id": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Bot Id"
          },
          "pinned": {
            "title": "Pinned",
            "type": "boolean"
          },
          "preview": {
            "title": "Preview",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "thread_id": {
            "minLength": 1,
            "title": "Thread Id",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "unread": {
            "title": "Unread",
            "type": "boolean"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          },
          "workspace_id": {
            "minLength": 1,
            "title": "Workspace Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "workspace_id",
          "name",
          "title",
          "description",
          "instructions",
          "color",
          "notify_on_finish",
          "pinned",
          "archived_at",
          "unread",
          "parent_bot_id",
          "thread_id",
          "preview",
          "status",
          "computer_mode",
          "updated_at",
          "created_at"
        ],
        "title": "Bot",
        "type": "object"
      },
      "BotList": {
        "properties": {
          "bots": {
            "items": {
              "$ref": "#/components/schemas/Bot"
            },
            "title": "Bots",
            "type": "array"
          }
        },
        "required": [
          "bots"
        ],
        "title": "BotList",
        "type": "object"
      },
...
      "CreateBotInput": {
        "properties": {
          "color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Color"
          },
          "computer_mode": {
            "default": "team",
            "enum": [
              "team",
              "dedicated"
            ],
            "title": "Computer Mode",
            "type": "string"
          },
          "description": {
            "default": "",
            "maxLength": 4000,
            "title": "Description",
            "type": "string"
          },
          "instructions": {
            "default": "",
            "maxLength": 20000,
            "title": "Instructions",
            "type": "string"
          },
          "name": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "notify_on_finish": {
            "default": true,
            "title": "Notify On Finish",
            "type": "boolean"
          },
          "title": {
            "default": "",
            "maxLength": 160,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "CreateBotInput",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OkResponse": {
        "properties": {
          "ok": {
            "default": true,
            "title": "Ok",
            "type": "boolean"
          }
        },
        "title": "OkResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Artek Buddy",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/bots": {
      "get": {
        "operationId": "list_bots_v1_bots_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BotList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Bots"
      },
      "post": {
        "operationId": "create_bot_v1_bots_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBotInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Bot"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Bot"
      }
    },
    "/v1/bots/archived": {
      "get": {
        "operationId": "list_archived_bots_v1_bots_archived_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BotList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Archived Bots"
      }
    },
    "/v1/bots/{bot_id}": {
      "delete": {
        "operationId": "remove_bot_v1_bots__bot_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "bot_id",
            "required": true,
            "schema": {
              "title": "Bot Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "delete_memories",
            "required": false,
            "schema": {
              "default": false,
              "title": "Delete Memories",
              "type": "boolean"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Remove Bot"
      },
      "get": {
        "operationId": "get_bot_v1_bots__bot_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "bot_id",
            "required": true,
            "schema": {
              "title": "Bot Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Bot"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Bot"
      },
      "patch": {
        "operationId": "update_bot_v1_bots__bot_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "bot_id",
            "required": true,
            "schema": {
              "title": "Bot Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBotInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Bot"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Bot"
      }
    },
    "/v1/bots/{bot_id}/archive": {
      "post": {
        "operationId": "archive_bot_v1_bots__bot_id__archive_post",
        "parameters": [
          {
            "in": "path",
            "name": "bot_id",
            "required": true,
            "schema": {
              "title": "Bot Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Archive Bot"
      }
    },
    "/v1/bots/{bot_id}/computer": {
      "post": {
        "operationId": "set_bot_computer_v1_bots__bot_id__computer_post",
        "parameters": [
          {
            "in": "path",
            "name": "bot_id",
            "required": true,
            "schema": {
              "title": "Bot Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetComputerInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Bot"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Bot Computer"
      }
    },
    "/v1/bots/{bot_id}/duplicate": {
      "post": {
        "operationId": "duplicate_bot_v1_bots__bot_id__duplicate_post",
        "parameters": [
          {
            "in": "path",
            "name": "bot_id",
            "required": true,
            "schema": {
              "title": "Bot Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Bot"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Duplicate Bot"
      }
    },
    "/v1/bots/{bot_id}/restore": {
      "post": {
        "operationId": "restore_bot_v1_bots__bot_id__restore_post",
        "parameters": [
          {
            "in": "path",
            "name": "bot_id",
            "required": true,
            "schema": {
              "title": "Bot Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Restore Bot"
      }
    },
    "/v1/bots/{bot_id}/subagents": {
      "get": {
        "operationId": "list_subagents_v1_bots__bot_id__subagents_get",
        "parameters": [
          {
            "in": "path",
            "name": "bot_id",
            "required": true,
            "schema": {
              "title": "Bot Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubagentList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Subagents"
      }
    },
    "/v1/bots/{bot_id}/subagents/{subagent_id}/restart": {
      "post": {
        "operationId": "restart_subagent_v1_bots__bot_id__subagents__subagent_id__restart_post",
        "parameters": [
          {
            "in": "path",
            "name": "bot_id",
            "required": true,
            "schema": {
              "title": "Bot Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "subagent_id",
            "required": true,
            "schema": {
              "title": "Subagent Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Subagent"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Restart Subagent"
      }
    },
    "/v1/bots/{bot_id}/subagents/{subagent_id}/steer": {
      "post": {
        "operationId": "steer_subagent_v1_bots__bot_id__subagents__subagent_id__steer_post",
        "parameters": [
          {
            "in": "path",
            "name": "bot_id",
            "required": true,
            "schema": {
              "title": "Bot Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "subagent_id",
            "required": true,
            "schema": {
              "title": "Subagent Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SteerSubagentInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Subagent"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Steer Subagent"
      }
    },
    "/v1/bots/{bot_id}/subagents/{subagent_id}/stop": {
      "post": {
        "operationId": "stop_subagent_v1_bots__bot_id__subagents__subagent_id__stop_post",
        "parameters": [
          {
            "in": "path",
            "name": "bot_id",
            "required": true,
            "schema": {
              "title": "Bot Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "subagent_id",
            "required": true,
            "schema": {
              "title": "Subagent Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Subagent"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stop Subagent"
      }
    },
...
  }
}